zeroguard_client/requirements.txt
flask==3.0.3
flask-sqlalchemy==3.1.1
flask-cors==4.0.1
flask-limiter==3.7.0
python-dotenv==1.0.1
PyJWT==2.8.0
cryptography==42.0.8
requests==2.31.0
groq==0.9.0
